+++ src/process/mod.rs
@@ -4,5 +4,6 @@ pub mod backfill;
 pub mod daemon;
 pub mod official_games;
 pub mod official_stats;
+pub mod shifts;
 pub mod status;
 pub mod sync;